RR-1791 - Move strengths middleware so that it can be re-used by other routes
