+++ server/routes/profile/overview/index.ts
@@ -3,8 +3,8 @@ import { Services } from '../../../services'
 import OverviewController from './overviewController'
 import asyncMiddleware from '../../../middleware/asyncMiddleware'
 import retrieveConditions from '../middleware/retrieveConditions'
-import retrieveStrengths from '../middleware/retrieveStrengths'
-import retrieveAlnScreeners from '../middleware/retrieveAlnScreeners'
+import retrieveStrengths from '../../middleware/retrieveStrengths'
+import retrieveAlnScreeners from '../../middleware/retrieveAlnScreeners'
 import retrieveCurrentChallenges from '../middleware/retrieveCurrentChallenges'
 import retrieveCuriousAlnAndLddAssessments from '../middleware/retrieveCuriousAlnAndLddAssessments'
 import retrievePrisonsLookup from '../../middleware/retrievePrisonsLookup'
