+++ server/routes/profile/strengths/index.ts
@@ -2,8 +2,8 @@ import { Router } from 'express'
 import asyncMiddleware from '../../../middleware/asyncMiddleware'
 import StrengthsController from './strengthsController'
 import { Services } from '../../../services'
-import retrieveStrengths from '../middleware/retrieveStrengths'
-import retrieveAlnScreeners from '../middleware/retrieveAlnScreeners'
+import retrieveStrengths from '../../middleware/retrieveStrengths'
+import retrieveAlnScreeners from '../../middleware/retrieveAlnScreeners'
 import retrievePrisonsLookup from '../../middleware/retrievePrisonsLookup'
 import retrieveEducationSupportPlanLifecycleStatus from '../middleware/retrieveEducationSupportPlanLifecycleStatus'
